# backend/openapi/v0.1.0.yaml
openapi: 3.0.3
info:
  title: Points are Bad REST API
  description: |-
    This is the OpenAPI configuration for the Points are Bad REST API. For the source, code please see the [GitHub](https://github.com/dangwn/points-are-bad).
#   termsOfService: 
#   contact:
#     email: 
  license:
    name: Points are Bad License
    url: https://github.com/dangwn/points-are-bad/LICENSE
  version: 0.1.0
...
servers:
  - url: https://api.pointsarebad.com
...
paths:
...
  /user/:
    get:
      tags:
        - user
      summary: Get current user
      description: Gets the current user's information
      deprecated: true
      responses:
        '202':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionUser'
        '401':
          description: Could not authenticate user
      security:
        - api_key: []
    post:
      tags:
        - user
      summary: Create new user
      description: Creates a new user with validation token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '400':
          description: Could not create new user
        '401':
          description: Could not create and assign tokens
    delete:
      tags:
        - user
      summary: Delete user
      description: Delete the current user
      responses:
        '204':
          description: User deleted
        '401':
          description: Could not authenticate user
      security:
        - api_key: []
  /user/username/:
    put:
      tags:
        - user
      summary: Edit username
      description: Edit the current user's username
      requestBody:
        content:
          application/json:
            schema:
              properties:
                username:
                  type: string
                  example: string
          application/x-www-form-urlencoded:
            schema:
              properties:
                username:
                  type: string
                  example: string
      responses:
        '202':
          description: Username updated
        '400':
          description: Could not update username
        '401':
          description: Could not authenticate user
      security:
        - api_key: []
  /user/password/:
    put:
      tags:
        - user
      summary: Edit password
      description: Edit the current user's password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPassword'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewPassword'
      responses:
        '202':
          description: Password updated
        '400':
          description: Could not update password
        '401':
          description: Could not authenticate user
      security:
        - api_key: []
components:
  schemas:
...
    NewPassword:
      type: object
      properties:
        old_password:
          type: string
          example: string
        new_password:
          type: string
          example: string
    NewUser:
      type: object
      properties:
        token: 
          type: string
          example: string
        username:
          type: string
          example: string
        password:
          type: string
          example: string
    SessionUser:
      type: object
      properties:
        username:
          type: string
          example: string
        is_admin:
          type: boolean
          example: false
    Token:
      type: object
      properties:
        access_token:
          type: string
          example: string
        token_type: 
          type: string
          example: string       
...
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header